fix: jsx pattern regression
