+++ packages/parser/__tests__/jsx-pattern.test.ts
@@ -1,7 +1,7 @@
-import { describe, test, expect } from 'vitest'
-import { patternParser } from './fixture'
+import { describe, expect, test } from 'vitest'
+import { parseAndExtract } from './fixture'
 
-describe('pattern jsx', () => {
+describe('jsx pattern', () => {
   test('should extract', () => {
     const code = `
        import { Stack } from "styled-system/jsx"
@@ -15,51 +15,100 @@ describe('pattern jsx', () => {
        }
      `
 
-    expect(patternParser(code)).toMatchInlineSnapshot(`
-      Map {
-        "Stack" => Set {
-          {
-            "box": {
-              "column": 16,
-              "line": 7,
-              "node": "JsxOpeningElement",
-              "type": "map",
-              "value": Map {
-                "align" => {
-                  "column": 36,
-                  "line": 7,
-                  "node": "StringLiteral",
-                  "type": "literal",
-                  "value": "center",
-                },
-                "marginTop" => {
-                  "column": 55,
-                  "line": 7,
-                  "node": "StringLiteral",
-                  "type": "literal",
-                  "value": "40px",
-                },
-                "marginBottom" => {
-                  "column": 75,
-                  "line": 7,
-                  "node": "StringLiteral",
-                  "type": "literal",
-                  "value": "42px",
-                },
-              },
+    const result = parseAndExtract(code)
+    expect(result.json).toMatchInlineSnapshot(`
+      [
+        {
+          "data": [
+            {
+              "align": "center",
+              "marginBottom": "42px",
+              "marginTop": "40px",
             },
-            "data": [
-              {
-                "align": "center",
-                "marginBottom": "42px",
-                "marginTop": "40px",
-              },
-            ],
-            "name": "Stack",
-            "type": "jsx-pattern",
+          ],
+          "name": "Stack",
+          "type": "jsx-pattern",
+        },
+      ]
+    `)
+  })
+
+  test('custom jsx', () => {
+    const code = `
+    <Form>
+        <Form.Group gap="4">
+          <p>Test</p>
+          <p>Test</p>
+        </Form.Group>
+        <Form.Action gap="2">Action</Form.Action>
+      </Form>
+    `
+
+    const result = parseAndExtract(code, {
+      patterns: {
+        extend: {
+          grid: {
+            jsx: ['Form.Group', 'Grid'],
+          },
+          stack: {
+            jsx: ['Form.Action', 'Stack'],
           },
         },
+      },
+    })
+
+    expect(result.json).toMatchInlineSnapshot(`
+      [
+        {
+          "data": [
+            {},
+          ],
+          "name": "Form",
+          "type": "jsx",
+        },
+        {
+          "data": [
+            {
+              "gap": "4",
+            },
+          ],
+          "name": "Form.Group",
+          "type": "jsx-pattern",
+        },
+        {
+          "data": [
+            {
+              "gap": "2",
+            },
+          ],
+          "name": "Form.Action",
+          "type": "jsx-pattern",
+        },
+      ]
+    `)
+
+    expect(result.css).toMatchInlineSnapshot(`
+      "@layer utilities {
+        .d_grid {
+          display: grid;
       }
+
+        .gap_4 {
+          gap: var(--spacing-4);
+      }
+
+        .d_flex {
+          display: flex;
+      }
+
+        .gap_2 {
+          gap: var(--spacing-2);
+      }
+
+        .flex_column {
+          flex-direction: column;
+      }
+      }"
     `)
   })
 })
