+++ packages/parser/__tests__/jsx-recipe.test.ts
@@ -1,5 +1,5 @@
 import { describe, test, expect } from 'vitest'
-import { jsxRecipeParser } from './fixture'
+import { jsxRecipeParser, parseAndExtract } from './fixture'
 
 describe('recipe jsx', () => {
   test('should extract', () => {
@@ -164,6 +164,95 @@ describe('recipe jsx', () => {
           },
         },
       }
+    `)
+  })
+
+  test('jsx recipes matching with dot component that are not from a namespace', () => {
+    const code = `
+    const App = () => {
+      return (
+        <Tabs.Root defaultValue="button">
+          <Tabs.List>
+            <Tabs.Trigger value="button">Button</Tabs.Trigger>
+            <Tabs.Trigger value="radio">Radio Group</Tabs.Trigger>
+            <Tabs.Trigger value="slider">Slider</Tabs.Trigger>
+            <Tabs.Indicator />
+          </Tabs.List>
+      </Tabs.Root>)
+    }
+     `
+    const result = parseAndExtract(code, {
+      theme: {
+        extend: {
+          slotRecipes: {
+            tabs: {
+              className: 'tabs',
+              slots: ['root', 'list', 'trigger', 'content', 'indicator'],
+              base: {
+                root: {
+                  display: 'flex',
+                },
+              },
+            },
+          },
+        },
+      },
+    })
+    expect(result.json).toMatchInlineSnapshot(`
+      [
+        {
+          "data": [
+            {},
+          ],
+          "name": "Tabs.Root",
+          "type": "jsx-recipe",
+        },
+        {
+          "data": [
+            {},
+          ],
+          "name": "Tabs.List",
+          "type": "jsx-recipe",
+        },
+        {
+          "data": [
+            {},
+          ],
+          "name": "Tabs.Trigger",
+          "type": "jsx-recipe",
+        },
+        {
+          "data": [
+            {},
+          ],
+          "name": "Tabs.Trigger",
+          "type": "jsx-recipe",
+        },
+        {
+          "data": [
+            {},
+          ],
+          "name": "Tabs.Trigger",
+          "type": "jsx-recipe",
+        },
+        {
+          "data": [
+            {},
+          ],
+          "name": "Tabs.Indicator",
+          "type": "jsx-recipe",
+        },
+      ]
+    `)
+
+    expect(result.css).toMatchInlineSnapshot(`
+      "@layer recipes.slots {
+        @layer _base {
+          .tabs__root {
+            display: flex;
+      }
+          }
+      }"
     `)
   })
 })
