+++ packages/parser/__tests__/output.test.ts
@@ -2942,140 +2942,6 @@ describe('extract to css output pipeline', () => {
     `)
   })
 
-  test('recipe.staticCss', () => {
-    const { ctx } = parseAndExtract('', {
-      theme: {
-        extend: {
-          recipes: {
-            textStyle: {
-              staticCss: [{ size: ['h1'] }],
-            },
-          },
-        },
-      },
-    })
-
-    const sheet = ctx.createSheet()
-    ctx.appendCssOfType('static', sheet)
-    const css = ctx.getCss(sheet)
-
-    expect(css).toMatchInlineSnapshot(`
-      "@layer recipes {
-        @layer _base {
-          .textStyle > :not([hidden]) ~ :not([hidden]) {
-            border-inline-start-width: 20px;
-            border-inline-end-width: 0px;
-      }
-
-          .textStyle {
-            font-family: var(--fonts-mono);
-      }
-      }
-
-        .textStyle--size_h1 {
-          font-size: 5rem;
-          line-height: 1em;
-          font-weight: 800;
-      }
-      }"
-    `)
-  })
-
-  test('staticCss.recipe base', () => {
-    const { ctx } = parseAndExtract('', {
-      staticCss: {
-        recipes: {
-          tooltipStyle: [],
-        },
-      },
-    })
-
-    const sheet = ctx.createSheet()
-    ctx.appendCssOfType('static', sheet)
-    const css = ctx.getCss(sheet)
-
-    expect(css).toMatchInlineSnapshot(`
-      "@layer recipes {
-        @layer _base {
-          [data-theme=dark] .tooltipStyle[data-tooltip],.dark .tooltipStyle[data-tooltip],.tooltipStyle[data-tooltip].dark,.tooltipStyle[data-tooltip][data-theme=dark],[data-theme=dark] .tooltipStyle [data-tooltip],.dark .tooltipStyle [data-tooltip],.tooltipStyle [data-tooltip].dark,.tooltipStyle [data-tooltip][data-theme=dark] {
-            color: red;
-      }
-          }
-      }"
-    `)
-  })
-
-  test('slotRecipes.staticCss', () => {
-    const { ctx } = parseAndExtract('', {
-      theme: {
-        extend: {
-          slotRecipes: {
-            someRecipe: {
-              staticCss: [{ size: ['sm'] }],
-              className: 'button',
-              slots: ['container', 'icon'],
-              base: {
-                container: {
-                  fontFamily: 'mono',
-                },
-                icon: {
-                  fontSize: '1.5rem',
-                },
-              },
-              variants: {
-                size: {
-                  sm: {
-                    container: {
-                      fontSize: '5rem',
-                      lineHeight: '1em',
-                    },
-                    icon: {
-                      fontSize: '2rem',
-                    },
-                  },
-
-                  md: {
-                    container: {
-                      fontSize: '3rem',
-                      lineHeight: '1.2em',
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
-    })
-
-    const sheet = ctx.createSheet()
-    ctx.appendCssOfType('static', sheet)
-    const css = ctx.getCss(sheet)
-
-    expect(css).toMatchInlineSnapshot(`
-      "@layer recipes.slots {
-        @layer _base {
-          .button__container {
-            font-family: var(--fonts-mono);
-      }
-
-          .button__icon {
-            font-size: 1.5rem;
-      }
-      }
-
-        .button__container--size_sm {
-          font-size: 5rem;
-          line-height: 1em;
-      }
-
-        .button__icon--size_sm {
-          font-size: 2rem;
-      }
-      }"
-    `)
-  })
-
   test('recipe issue', () => {
     const code = `
     import { css } from 'styled-system/css';
@@ -3186,114 +3052,6 @@ describe('extract to css output pipeline', () => {
       "@layer recipes {
         .card--rounded_true {
           border-radius: 0.375rem;
-      }
-      }"
-    `)
-  })
-
-  test('sva with unresolvable slots', () => {
-    const code = `
-    import { sva } from 'styled-system/css'
-    import { slots } from './slots'
-
-    const card = sva({
-      slots,
-      base: {
-        root: {
-          p: '6',
-          m: '4',
-          w: 'md',
-          boxShadow: 'md',
-          borderRadius: 'md',
-          _dark: { bg: '#262626', color: 'white' },
-        },
-        content: {
-          textStyle: 'lg',
-        },
-        title: {
-          textStyle: 'xl',
-          fontWeight: 'semibold',
-          pb: '2',
-        },
-      },
-    })
-     `
-    const result = parseAndExtract(code)
-    expect(result.json).toMatchInlineSnapshot(`
-      [
-        {
-          "data": [
-            {
-              "base": {
-                "content": {
-                  "textStyle": "lg",
-                },
-                "root": {
-                  "_dark": {
-                    "bg": "#262626",
-                    "color": "white",
-                  },
-                  "borderRadius": "md",
-                  "boxShadow": "md",
-                  "m": "4",
-                  "p": "6",
-                  "w": "md",
-                },
-                "title": {
-                  "fontWeight": "semibold",
-                  "pb": "2",
-                  "textStyle": "xl",
-                },
-              },
-              "slots": [
-                "root",
-                "content",
-                "title",
-              ],
-            },
-          ],
-          "name": "sva",
-          "type": "sva",
-        },
-      ]
-    `)
-
-    expect(result.css).toMatchInlineSnapshot(`
-      "@layer utilities {
-        .p_6 {
-          padding: var(--spacing-6);
-      }
-
-        .m_4 {
-          margin: var(--spacing-4);
-      }
-
-        .w_md {
-          width: var(--sizes-md);
-      }
-
-        .shadow_md {
-          box-shadow: var(--shadows-md);
-      }
-
-        .rounded_md {
-          border-radius: var(--radii-md);
-      }
-
-        .fw_semibold {
-          font-weight: var(--font-weights-semibold);
-      }
-
-        .pb_2 {
-          padding-bottom: var(--spacing-2);
-      }
-
-        [data-theme=dark] .dark\\:bg_\\#262626,.dark .dark\\:bg_\\#262626,.dark\\:bg_\\#262626.dark,.dark\\:bg_\\#262626[data-theme=dark] {
-          background: #262626;
-      }
-
-        [data-theme=dark] .dark\\:text_white,.dark .dark\\:text_white,.dark\\:text_white.dark,.dark\\:text_white[data-theme=dark] {
-          color: var(--colors-white);
       }
       }"
     `)
@@ -3680,297 +3438,6 @@ describe('extract to css output pipeline', () => {
     `)
   })
 
-  test('TS namespaces - patterns', () => {
-    const code = `
-    import * as p from "styled-system/patterns"
-
-    p.stack({ mt: "40px" })
-     `
-    const result = parseAndExtract(code)
-    expect(result.json).toMatchInlineSnapshot(`
-      [
-        {
-          "data": [
-            {
-              "mt": "40px",
-            },
-          ],
-          "name": "stack",
-          "type": "pattern",
-        },
-      ]
-    `)
-
-    expect(result.css).toMatchInlineSnapshot(`
-      "@layer utilities {
-        .d_flex {
-          display: flex;
-      }
-
-        .gap_10px {
-          gap: 10px;
-      }
-
-        .flex_column {
-          flex-direction: column;
-      }
-
-        .mt_40px {
-          margin-top: 40px;
-      }
-      }"
-    `)
-  })
-
-  test('TS namespaces - recipes', () => {
-    const code = `
-    import * as recipes from "styled-system/recipes"
-
-    recipes.cardStyle({ rounded: true })
-     `
-    const result = parseAndExtract(code)
-    expect(result.json).toMatchInlineSnapshot(`
-      [
-        {
-          "data": [
-            {
-              "rounded": true,
-            },
-          ],
-          "name": "cardStyle",
-          "type": "recipe",
-        },
-      ]
-    `)
-
-    expect(result.css).toMatchInlineSnapshot(`
-      "@layer recipes {
-        .card--rounded_true {
-          border-radius: 0.375rem;
-      }
-      }"
-    `)
-  })
-
-  test('TS namespaces - css', () => {
-    const code = `
-    import * as panda from "styled-system/css"
-
-    panda.css({ color: "red" })
-    panda.cva({ base: { color: "blue" } })
-    panda.sva({ base: { root: { color: "green" } } })
-     `
-    const result = parseAndExtract(code)
-    expect(result.json).toMatchInlineSnapshot(`
-      [
-        {
-          "data": [
-            {
-              "color": "red",
-            },
-          ],
-          "name": "css",
-          "type": "css",
-        },
-        {
-          "data": [
-            {
-              "base": {
-                "color": "blue",
-              },
-            },
-          ],
-          "name": "cva",
-          "type": "cva",
-        },
-        {
-          "data": [
-            {
-              "base": {
-                "root": {
-                  "color": "green",
-                },
-              },
-              "slots": [
-                "root",
-              ],
-            },
-          ],
-          "name": "sva",
-          "type": "sva",
-        },
-      ]
-    `)
-
-    expect(result.css).toMatchInlineSnapshot(`
-      "@layer utilities {
-        .text_red {
-          color: red;
-      }
-
-        .text_blue {
-          color: blue;
-      }
-
-        .text_green {
-          color: green;
-      }
-      }"
-    `)
-  })
-
-  test('TS namespaces - jsx', () => {
-    const code = `
-    import * as JSX from "styled-system/jsx"
-
-    const App = () => {
-      return <>
-      <JSX.styled.div color="red" />
-      <JSX.Stack color="blue" />
-      <JSX.Grid color="green" />
-      </>
-    }
-     `
-    const result = parseAndExtract(code)
-    expect(result.json).toMatchInlineSnapshot(`
-      [
-        {
-          "data": [
-            {
-              "color": "red",
-            },
-          ],
-          "name": "JSX.styled.div",
-          "type": "jsx-factory",
-        },
-        {
-          "data": [
-            {
-              "color": "blue",
-            },
-          ],
-          "name": "Stack",
-          "type": "jsx-pattern",
-        },
-        {
-          "data": [
-            {
-              "color": "green",
-            },
-          ],
-          "name": "Grid",
-          "type": "jsx-pattern",
-        },
-      ]
-    `)
-
-    expect(result.css).toMatchInlineSnapshot(`
-      "@layer utilities {
-        .text_red {
-          color: red;
-      }
-
-        .d_flex {
-          display: flex;
-      }
-
-        .gap_10px {
-          gap: 10px;
-      }
-
-        .text_blue {
-          color: blue;
-      }
-
-        .d_grid {
-          display: grid;
-      }
-
-        .text_green {
-          color: green;
-      }
-
-        .flex_column {
-          flex-direction: column;
-      }
-      }"
-    `)
-  })
-
-  test('TS namespaces - ignore not from panda', () => {
-    const code = `
-    import * as panda from "not-panda"
-
-    panda.css({ color: "red" })
-    panda.cva({ base: { color: "blue" } })
-    panda.sva({ base: { root: { color: "green" } } })
-     `
-    const result = parseAndExtract(code)
-    expect(result.json).toMatchInlineSnapshot(`[]`)
-
-    expect(result.css).toMatchInlineSnapshot(`""`)
-  })
-
-  test('TS namespaces - JSX factory', () => {
-    const code = `
-    import * as pandaJsx from '../styled-system/jsx';
-
-    pandaJsx.styled('div', { base: { color: 'red' } })
-    const App = () => <pandaJsx.styled.span color="blue">Hello</pandaJsx.styled.span>
-     `
-    const result = parseAndExtract(code)
-    expect(result.json).toMatchInlineSnapshot(`
-      [
-        {
-          "data": [
-            {
-              "base": {
-                "color": "red",
-              },
-            },
-          ],
-          "name": "pandaJsx.styled",
-          "type": "cva",
-        },
-        {
-          "data": [
-            {
-              "color": "blue",
-            },
-          ],
-          "name": "pandaJsx.styled.span",
-          "type": "jsx-factory",
-        },
-      ]
-    `)
-
-    expect(result.css).toMatchInlineSnapshot(`
-      "@layer utilities {
-        .text_red {
-          color: red;
-      }
-
-        .text_blue {
-          color: blue;
-      }
-      }"
-    `)
-  })
-
-  test('TS namespaces - ignore not from panda', () => {
-    const code = `
-    import * as panda from "not-panda"
-
-    panda.css({ color: "red" })
-    panda.cva({ base: { color: "blue" } })
-    panda.sva({ base: { root: { color: "green" } } })
-     `
-    const result = parseAndExtract(code)
-    expect(result.json).toMatchInlineSnapshot(`[]`)
-
-    expect(result.css).toMatchInlineSnapshot(`""`)
-  })
-
   test('flat and nested object on same key', () => {
     const code = `
     const className = css({
@@ -4037,95 +3504,6 @@ describe('extract to css output pipeline', () => {
         .border_black\\.20 {
           border-color: var(--colors-black-20);
       }
-      }"
-    `)
-  })
-
-  test('JSX recipes matching with dot component that are not from a namespace', () => {
-    const code = `
-    const App = () => {
-      return (
-        <Tabs.Root defaultValue="button">
-          <Tabs.List>
-            <Tabs.Trigger value="button">Button</Tabs.Trigger>
-            <Tabs.Trigger value="radio">Radio Group</Tabs.Trigger>
-            <Tabs.Trigger value="slider">Slider</Tabs.Trigger>
-            <Tabs.Indicator />
-          </Tabs.List>
-      </Tabs.Root>)
-    }
-     `
-    const result = parseAndExtract(code, {
-      theme: {
-        extend: {
-          slotRecipes: {
-            tabs: {
-              className: 'tabs',
-              slots: ['root', 'list', 'trigger', 'content', 'indicator'],
-              base: {
-                root: {
-                  display: 'flex',
-                },
-              },
-            },
-          },
-        },
-      },
-    })
-    expect(result.json).toMatchInlineSnapshot(`
-      [
-        {
-          "data": [
-            {},
-          ],
-          "name": "Tabs.Root",
-          "type": "jsx-recipe",
-        },
-        {
-          "data": [
-            {},
-          ],
-          "name": "Tabs.List",
-          "type": "jsx-recipe",
-        },
-        {
-          "data": [
-            {},
-          ],
-          "name": "Tabs.Trigger",
-          "type": "jsx-recipe",
-        },
-        {
-          "data": [
-            {},
-          ],
-          "name": "Tabs.Trigger",
-          "type": "jsx-recipe",
-        },
-        {
-          "data": [
-            {},
-          ],
-          "name": "Tabs.Trigger",
-          "type": "jsx-recipe",
-        },
-        {
-          "data": [
-            {},
-          ],
-          "name": "Tabs.Indicator",
-          "type": "jsx-recipe",
-        },
-      ]
-    `)
-
-    expect(result.css).toMatchInlineSnapshot(`
-      "@layer recipes.slots {
-        @layer _base {
-          .tabs__root {
-            display: flex;
-      }
-          }
       }"
     `)
   })
